app/pyproject.toml
[project]
name = "twitter-collector"
version = "0.1.0"
description = "Twitter/X collection pipeline backed by twitter-cli"
requires-python = ">=3.12"
dependencies = [
    "pipeline-common",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml",
]

[tool.uv.sources]
pipeline-common = { path = "../common" }

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = ["pytest>=8.0", "pytest-cov>=5.0"]
